multi: add StopDaemon method to Lit
In this commit, a new Proxy service is added with a StopDaemon method.
This method can be used to stop the Litd service. This will be useful in
remote-mode itests where we want to restart Litd without also restarting
LND. It also provides a nice way of shutting down Litd in remote-mode.
A Ping method is also added to the service. The reason for this method
is so that access to the new Proxy service can be tested in the itests
withouth actually shutting down Litd.